+++ Source/CesiumRuntime/Public/Cesium3DTileset.h
@@ -244,16 +244,13 @@ public:
   float CulledScreenSpaceError = 64.0;
 
   /**
-   *  TODO: should this be under rendering
+   * Refreshes this tileset, ensuring that all materials and other settings are
+   * applied. It is not usually necessary to invoke this, but when
+   * behind-the-scenes changes are made and not reflected in the tileset, this
+   * function can help.
    */
-  UFUNCTION(CallInEditor, Category = "Cesium|Rendering")
+  UFUNCTION(CallInEditor, BlueprintCallable, Category = "Cesium")
   void RefreshTileset();
-
-  /**
-   *
-   */
-  UPROPERTY(EditAnywhere, Category = "Cesium|Rendering")
-  TArray<ACesiumCartographicSelection*> CartographicSelections;
 
   /**
    * Pauses level-of-detail and culling updates of this tileset.
@@ -398,20 +395,6 @@ private:
       Category = "Cesium|Rendering")
   UMaterialInterface* WaterMaterial = nullptr;
 
-  /**
-   * A custom Material to use to render this tileset in areas using opacity
-   * masks, in order to implement custom visual effects.
-   *
-   * The custom material should generally be created by copying the
-   * "M_CesiumDefaultMasked" material and customizing it as desired.
-   */
-  UPROPERTY(
-      EditAnywhere,
-      BlueprintGetter = GetOpacityMaskMaterial,
-      BlueprintSetter = SetOpacityMaskMaterial,
-      Category = "Cesium|Rendering")
-  UMaterialInterface* OpacityMaskMaterial = nullptr;
-
 protected:
   UPROPERTY()
   FString PlatformName;
@@ -464,14 +447,6 @@ public:
 
   UFUNCTION(BlueprintSetter, Category = "Cesium|Rendering")
   void SetWaterMaterial(UMaterialInterface* InMaterial);
-
-  UFUNCTION(BlueprintGetter, Category = "Cesium|Rendering")
-  UMaterialInterface* GetOpacityMaskMaterial() const {
-    return OpacityMaskMaterial;
-  }
-
-  UFUNCTION(BlueprintSetter, Category = "Cesium|Rendering")
-  void SetOpacityMaskMaterial(UMaterialInterface* InMaterial);
 
   UFUNCTION(BlueprintCallable, Category = "Cesium|Rendering")
   void PlayMovieSequencer();